16 спринт MVVM

+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackDto.kt
@@ -3,7 +3,7 @@ package com.example.playlistmaker.data.dto
 data class TrackDto(
     val trackName: String,
     val artistName: String,
-    val trackTimeMillis: Int,
+    val trackTimeMillis: Long,
     val artworkUrl100: String,
     val trackId: Int,
     var collectionName: String,
